backend/requirements.txt
redis
aioredis
python-jose
slowapi
structlog
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
sqlalchemy
psycopg2
asyncio
alembic
pytest
httpx